Split benchmark for reduction operators
